feat(file): 实现文件上传至MinIO功能

+++ src/main/java/anthony/tikax/domain/model/UploadFileDO.java
@@ -10,7 +10,7 @@ public class UploadFileDO {
 
     private String fileMd5;
     private String fileName;
-    private Integer totalSize;
+    private Long totalSize;
     private String fileType;
     private String extension;
     private String mimeType;
